Render internal links as relative links.
There were some locations where the site was linking to the listing
of plugin tags with an absolute ("/plugin/tags") path.  This fixes
those to render as relative paths.
This matches the rendering of all other internal links on the site,
and avoids broken links when deploying to sites rooted under a path
prefix (e.g. https://lektor.github.io/lektor-website/.)