Remove display names from email fields.
As it turns out, RFC6068, the current spec for `mailto:` URIs
wants plain _addr-spec_s (e.g. `localpart@domain`) in the URLs.
It does not allow display names in the addresses (e.g. `Joe Bloe
<localpart@domain>`) as the previous spec (RFC2638) did.

So here, we remove any display names from the email addresses
in the "Author-Email" (and "Maintainer-Email") field(s).